Ccommand-line arguments support for input, output and operations

+++ src/main/java/app/integrations/utils/FileHelper.java
@@ -3,8 +3,22 @@ package app.integrations.utils;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
+import java.nio.file.FileAlreadyExistsException;
 
 public class FileHelper {
+
+    public static File createFile(String absoluteFileName, boolean overwrite) throws IOException {
+        File file = new File(absoluteFileName);
+        file.getParentFile().mkdirs();
+        if (!overwrite && file.exists()) {
+            throw new FileAlreadyExistsException(absoluteFileName);
+        }
+        if (overwrite && file.exists()) {
+            file.delete();
+        }
+        file.createNewFile();
+        return file;
+    }
 
     public static File createFile(String foldername, String filename, boolean overwrite) throws IOException {
         File folder = new File(foldername);
@@ -15,6 +29,9 @@ public class FileHelper {
             throw new FileNotFoundException("'" + folder.getAbsolutePath() + "' is not a directory");
         }
         File file = new File(folder, filename);
+        if (!overwrite && file.exists()) {
+            throw new FileAlreadyExistsException("'" + filename + "' in '" + foldername + "'");
+        }
         if (overwrite && file.exists()) {
             file.delete();
         }
@@ -27,6 +44,14 @@ public class FileHelper {
         File file = new File(folder, filename);
         if (!file.exists() || file.isDirectory()) {
             throw new FileNotFoundException("'" + filename + "' in '" + foldername + "'");
+        }
+        return file;
+    }
+
+    public static File checkAndGetFile(String absoluteFileName) throws FileNotFoundException {
+        File file = new File(absoluteFileName);
+        if (!file.exists() || file.isDirectory()) {
+            throw new FileNotFoundException(absoluteFileName);
         }
         return file;
     }
